requirements.txt
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.0
fastapi==0.104.1
flower==2.0.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
humanize==4.9.0
idna==3.6
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
pika==1.3.2
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg-pool==3.2.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
Secweb==1.7.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
